Revisi Database

+++ server/models/data_pembiayaan_saat_ini.js
@@ -17,8 +17,8 @@ module.exports = (sequelize, DataTypes) => {
     Id_user: DataTypes.INTEGER,
     id_form_pembiayaan: DataTypes.INTEGER,
     pembiayaan_bank_lain: DataTypes.STRING,
-    jumlah_pembiayaan: DataTypes.INTEGER,
-    angsuran: DataTypes.INTEGER,
+    jumlah_pembiayaan: DataTypes.BIGINT,
+    angsuran: DataTypes.BIGINT,
     jatuh_tempo: DataTypes.INTEGER,
     jenis_pembiayaan: DataTypes.STRING,
     nama_kreditur: DataTypes.STRING
